requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
asttokens==3.0.1
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.5
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.124.4
Flask==3.1.2
fsspec==2026.2.0
genkit==0.4.0
genkit-plugin-google-genai==0.4.0
google-api-core==2.28.1
google-auth==2.48.0
google-cloud-aiplatform==1.126.1
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==3.5.0
google-crc32c==1.7.1
google-genai==1.63.0
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0

ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.1
lxml==6.0.2
mangum==0.21.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
numpy==1.26.4
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.7
packaging==24.2
parso==0.8.5
partial-json-parser==0.2.1.1.post6
pdf2image==1.17.0
pdfrw==0.4
pexpect==4.9.0
pillow==12.0.0
pinecone==8.0.0
pinecone-plugin-assistant==3.0.2
pinecone-plugin-interface==0.0.7
postgrest==2.28.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyiceberg==0.11.0
PyJWT==2.11.0
pyparsing==3.3.2
pypdf==5.2.0
pyroaring==1.0.3
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
python-pptx==1.0.2
realtime==2.28.0
requests==2.32.5
rich==14.3.3
rsa==4.9.1
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.50.0
storage3==2.28.0
StrEnum==0.4.15
strictyaml==1.7.3
structlog==25.5.0
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
tenacity==9.1.2
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
vertexai==1.43.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.5
xlsxwriter==3.2.9
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
youtube-transcript-api
yt-dlp
faster-whisper
deepgram-sdk==3.*